requirements.txt
icecream
requests
msgspec
surrealdb
python-dotenv
python-telegram-bot
